Hotel-Project
I'm a beginner with little bit of experience in C++, (Code is written in C++) . 
What I am trying to do is, have a hotel system that removes/adds customers to a hotel room. 
What's done or still working on:
A: Display Hotel Rooms 1 to 10 (Complete)
B: Have a function in place that adds customers (Complete)
C: Menu system & Error check such as user inputs an invalid in put (Complete)

The Problem: 
My array is not updating when I add customer. First time the code is run all the rooms are empty, than I add a customer and they select their desired room number. When the View rooms is called one room should be occupied by a customer. However, it is still showing that all the rooms are empty.

Thank for your help in advance